test/fixtures/wpt/performance-timeline/tentative/include-frames-originA-B-B.html

<!DOCTYPE html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src=/common/get-host-info.sub.js></script>

</head>

<body>
  <script>
    const loadChildFrame = () => {
      return new Promise(resolve => {

        const crossOriginChildFrame = document.createElement('iframe');

        crossOriginChildFrame.src = get_host_info().REMOTE_ORIGIN +
          "/performance-timeline/resources/include-frames-subframe.html?origin=" + get_host_info().REMOTE_ORIGIN;

        document.body.appendChild(crossOriginChildFrame)

        window.addEventListener("message", e => {
          resolve(e.data);
        });
      });
    }

    promise_test(async () => {
      performance.clearResourceTimings();

      // Load a child frame. The child frame upon loading would load a child frame of its own.
      childFrameEntrySize = await loadChildFrame();

      // Verify the number of performance entries in the child frame.
      // 4 for child, 2 for grandchild.
      assert_equals(childFrameEntrySize, 6, 'Child frame entries should be 6.');

      const entries = performance.getEntries({ includeChildFrames: true });

      // 3 entries for parent, 0 for child, 0 for grandchild.
      assert_equals(entries.length, 3, 'Total entries should be 3.');
    }, 'GetEntries of a document of origin A, its child frame of origin B and \
    its grandchild frame of origin B.');
  </script>
</body>
